backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.2.0
django-environ==0.11.2
drf-spectacular==0.27.1
PyJWT==2.8.0
